pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = "Task manager app"
authors = ["roaddust2 <roaddust2@yahoo.com>"]
readme = "README.md"
packages = [{include = "task_manager"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.7"
python-dotenv = "^0.21.1"
psycopg2-binary = "^2.9.5"
gunicorn = "^20.1.0"
whitenoise = "^6.3.0"
django-bootstrap4 = "^22.3"
django-filter = "^22.1"
dj-database-url = "^1.2.0"
rollbar = "^0.16.3"


[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"
pytest-django = "^4.5.2"
pytest-cov = "^4.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
